requirements.txt
streamlit>=1.37.0
google-generativeai>=0.7.2
requests>=2.32.3
duckduckgo-search>=6.1.3
trafilatura>=1.7.0
beautifulsoup4>=4.12.3
lxml>=5.3.0
ddgs>=1.0.7
faiss-cpu>=1.8.0
